Abstract
The term Ethernet Time Synchronization (ETS) in this white paper applies to both physical layer and
packet-based techniques that provide frequency and time-of-day synchronization across an Ethernet net-
work. This document describes the necessity of ETS and its applications within various market segments.
It then provides an overview of the techniques and protocols that provide the ETS function.